+++ builder/builders/compilers/definitions.ts
@@ -1,0 +1,12 @@
+import { ExportType } from './../utils';
+
+export interface BuilderResult {
+  path: string;
+  type: ExportType;
+  affectedFiles: string[];
+}
+
+export interface CompilerResult extends BuilderResult {
+  output: string;
+  file: string;
+}